Move to ttfinstrs.h all the prototypes of non-static ttfinstrs.c functions
