+++ fontforge/sfd.c
@@ -46,6 +46,7 @@
 #include "splineutil.h"
 #include "splineutil2.h"
 #include "tottfgpos.h"
+#include "ttfinstrs.h"
 #include "baseviews.h"
 #include "views.h"
 #include <gdraw.h>
