+++ fontforge/ttfinstrs.c
@@ -24,11 +24,13 @@
  * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
  * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+
+#include "ttfinstrs.h"
+
 #include "fontforgevw.h"
 #include <ustring.h>
 #include <utype.h>
 #include "ttf.h"
-#include "ttfinstrs.h"
 
 const char *ff_ttf_instrnames[] = {
     "SVTCA[y-axis]",
